\docType{data}
\name{example_pwp}
\alias{example_pwp}
\title{Example pairwise pi matrix used in a \code{sandwrm} analysis}
\usage{
example_pwp
}
\description{
Pairwise pi matrix from a simulated dataset to run an example
\code{sandwrm} analysis. Loads as a data.frame, will need to be converted into a matrix.
}
\keyword{datasets}
