\docType{package}
\name{sandwrm-package}
\alias{sandwrm-package}
\title{The 'sandwrm' package.}
\description{
A method for inferring Wright's neighborhood size and species-level diversity in continuous space by fitting the Wright-Malecot model to a plot of geographic distance against genetic distance. This package contains the function sandwrm(), which performs the analysis and outputs several plots of ease of visualization and fit diagnostics. See papers in the references below for additional details.
}
\references{
Z.B. Hancock, R.H. Toczydlowski, and G.S. Bradburd (2024) <doi: 10.1093/genetics/iyae094>.

Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
